Extract state object for record/replay modes
