+++ lib/impersonator/recording.rb
@@ -3,6 +3,21 @@
 module Impersonator
   # A recording is responsible for saving interactions at record time, and replaying them at
   # replay time.
+  #
+  # A recording is always in one of two states.
+  #
+  # * {RecordMode Record mode}
+  # * {ReplayMode Replay mode}
+  #
+  # The state objects are responsible of dealing with the recording logic, which happens in 3
+  # moments:
+  #
+  # * {#start}
+  # * {#invoke}
+  # * {#finish}
+  #
+  # @see RecordMode
+  # @see ReplayMode
   class Recording
     include HasLogger
 
@@ -15,44 +30,39 @@ module Impersonator
       @label = label
       @recordings_path = recordings_path
       @disabled = disabled
+
+      initialize_current_mode
     end
 
     # Start a recording/replay session
     def start
       logger.debug "Starting recording #{label}..."
-      if can_replay?
-        start_in_replay_mode
-      else
-        start_in_record_mode
-      end
+      current_mode.start
     end
 
+    # Handles the invocation of a given method on the impersonated object
+    #
+    # It will either record the interaction or replay it dependening on if there
+    # is a recording available or not
+    #
+    # @param [Object, Double] impersonated_object
+    # @param [MethodInvocation] method
+    # @param [Array<Object>] args
     def invoke(impersonated_object, method, args)
-      if replay_mode?
-        replay(method)
-      else
-        spiable_block = method&.block_spy&.block
-        impersonated_object.send(method.name, *args, &spiable_block).tap do |return_value|
-          record(method, return_value)
-        end
-      end
+      current_mode.invoke(impersonated_object, method, args)
     end
 
     # Finish a record/replay session.
     def finish
       logger.debug "Recording #{label} finished"
-      if record_mode?
-        finish_in_record_mode
-      else
-        finish_in_replay_mode
-      end
+      current_mode.finish
     end
 
     # Return whether it is currently at replay mode
     #
     # @return [Boolean]
     def replay_mode?
-      @replay_mode
+      @current_mode == replay_mode
     end
 
     # Return whether it is currently at record mode
@@ -64,90 +74,34 @@ module Impersonator
 
     private
 
+    attr_reader :current_mode
+
+    def initialize_current_mode
+      @current_mode = if can_replay?
+                        replay_mode
+                      else
+                        record_mode
+                      end
+    end
+
     def can_replay?
-      !@disabled && File.exist?(file_path)
+      !@disabled && File.exist?(recording_path)
     end
 
-    # Record a {MethodInvocation method invocation} with a given return value
-    # @param [Method] method
-    # @param [Object] return_value
-    def record(method, return_value)
-      method_invocation = MethodInvocation.new(method_instance: method, return_value: return_value)
-
-      @method_invocations << method_invocation
+    def record_mode
+      @record_mode ||= RecordMode.new(recording_path)
     end
 
-    # Replay a method invocation
-    # @param [Method] method
-    def replay(method)
-      method_invocation = @method_invocations.shift
-      unless method_invocation
-        raise Impersonator::Errors::MethodInvocationError, 'Unexpected method invocation received:'\
-              "#{method}"
-      end
-
-      validate_method_signature!(method, method_invocation.method_instance)
-      replay_block(method_invocation, method)
-
-      method_invocation.return_value
+    def replay_mode
+      @replay_mode ||= ReplayMode.new(recording_path)
     end
 
-    def replay_block(recorded_method_invocation, method_to_replay)
-      block_spy = recorded_method_invocation.method_instance.block_spy
-      block_spy&.block_invocations&.each do |block_invocation|
-        method_to_replay.block.call(*block_invocation.arguments)
-      end
-    end
-
-    def start_in_replay_mode
-      logger.debug 'Replay mode'
-      @replay_mode = true
-      @method_invocations = YAML.load_file(file_path)
-    end
-
-    def start_in_record_mode
-      logger.debug 'Recording mode'
-      @replay_mode = false
-      make_sure_recordings_dir_exists
-      @method_invocations = []
-    end
-
-    def finish_in_record_mode
-      File.open(file_path, 'w') do |file|
-        YAML.dump(@method_invocations, file)
-      end
-    end
-
-    def finish_in_replay_mode
-      unless @method_invocations.empty?
-        raise Impersonator::Errors::MethodInvocationError,
-              "Expecting #{@method_invocations.length} method invocations"\
-              " that didn't happen: #{@method_invocations.inspect}"
-      end
-    end
-
-    def file_path
+    def recording_path
       File.join(@recordings_path, "#{label_as_file_name}.yml")
     end
 
     def label_as_file_name
       label.downcase.gsub(/[\(\)\s \#:]/, '-').gsub(/[\-]+/, '-').gsub(/(^-)|(-$)/, '')
-    end
-
-    def make_sure_recordings_dir_exists
-      dirname = File.dirname(file_path)
-      FileUtils.mkdir_p(dirname) unless File.directory?(dirname)
-    end
-
-    def validate_method_signature!(expected_method, actual_method)
-      unless actual_method == expected_method
-        raise Impersonator::Errors::MethodInvocationError, <<~ERROR
-          Expecting:
-            #{expected_method}
-          But received:
-            #{actual_method}
-        ERROR
-      end
     end
   end
 end
